api/requirements.txt
alembic==0.9.6
bcrypt==3.1.4
click==8.0
coverage==4.4.2
eventlet==0.31.0
flask>=0.12.3
Flask-Bcrypt==0.7.1
Flask-Migrate==2.1.1
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.7.1
flask_cors==1.10.3
gunicorn==19.7.1
itsdangerous==0.24
Jinja2>=2.10.1
jsonschema==2.6.0
Mako==1.0.7
psycopg2-binary==2.8.5
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2017.3
selenium==3.8.1
six==1.11.0
SQLAlchemy>=1.3.0
pillow>=8.4.0
celery==5.2.1
redis==4.0.2
faker==10.0.0
Werkzeug==0.16.0
